Move REPL word definitions to separate file
In order to re-use REPL word definitions both in Emscripten and CLI
versions of the REPL, separate the definitions to their own file.
